now using never throw

+++ src/dal/customers.ts
@@ -1,4 +1,6 @@
 import { auth } from "@clerk/nextjs/server";
+import { errAsync, okAsync } from "neverthrow";
+import z from "zod";
 import {
   createCustomerDb,
   deleteCustomerDb,
@@ -25,39 +27,43 @@ export async function getCustomers(): Promise<Result<Customer[]>> {
   }
 }
 
-export async function createCustomerDal(
-  input: CreateCustomerInput,
-): Promise<Result<Customer>> {
+export async function createCustomerDal(input: CreateCustomerInput) {
   const { orgId } = await auth.protect();
 
   if (!orgId) {
-    return { data: null, error: "No org" };
+    return errAsync({ reason: "Not authorized" } as const);
   }
 
   const validation = CreateCustomerSchema.safeParse(input);
   if (!validation.success) {
-    return { data: null, error: validation.error.issues[0].message };
+    const errorTree = z.treeifyError(validation.error);
+
+    return errAsync({
+      reason: "Validation failed",
+      errors: errorTree,
+    } as const);
   }
 
   try {
-    const data = await createCustomerDb(input, orgId);
-    return { data, error: null };
+    const customer = await createCustomerDb(input, orgId);
+    return okAsync(customer);
   } catch (e: unknown) {
     console.error("Database Insert Error:", e);
-    return {
-      data: null,
-      error: "Failed to create customer. Please try again.",
-    };
+    return errAsync({ reason: "Db failed to create customer" } as const);
   }
 }
 
-export async function deleteCustomerDal(id: string): Promise<Result<Customer>> {
-  await auth.protect();
+export async function deleteCustomerDal(id: string) {
+  const { orgId } = await auth.protect();
+  if (!orgId) {
+    return errAsync({ reason: "Not authorized" } as const);
+  }
+
   try {
     const data = await deleteCustomerDb(id);
-    return { data, error: null };
+    return okAsync(data);
   } catch (e: unknown) {
     console.error("Database Delete Error:", e);
-    return { data: null, error: "Database Error" };
+    return errAsync({ reason: "Db failed to delete customer" } as const);
   }
 }
